requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.3
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.61.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyodbc==5.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.44.0
Werkzeug==3.1.4
wikipedia==1.4.0
